Split the Independent Alert Engine region out of Dashboard MainWindow
Extracted the entire `#region Independent Alert Engine` (CheckAllServerAlertsAsync,
EvaluateAlertConditionsAsync, SendDetectedAlertAsync, and the eight Build*Context
builders, including the #1173 failed-job watermark logic) verbatim into a new
MainWindow.AlertEngine.cs partial. Same class, same namespace; the alert state
fields stay in MainWindow.xaml.cs and remain accessible to the partial.

Pure partial-class move — behavior-neutral (main-file diff is deletion-only, the
region is byte-identical in the new file), no call-site changes. The live alert
path is unchanged.

MainWindow.xaml.cs 2691 -> 1528; MainWindow.AlertEngine.cs 1203. Solution builds
clean; Dashboard.Tests (491) pass.
